+++ bio-common/src/main/kotlin/bio/relatives/common/model/Region.kt
@@ -13,8 +13,8 @@ data class Region(
     override val sequence: String,
     override val chromosome: String,
     override val gene: String,
-    override val start: Long,
-    override val end: Long
+    override val start: Int,
+    override val end: Int
 ) : NucleotideSequenceAware, ChromosomeAware, ChromosomePositionAware, IdentifierAware {
     init {
         validateChrom(this)
